extracted parsing tests and cleaned up segmentcoll tests
